# agent-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Intelli Agent Gateway
  version: 0.5.0
  description: |
    Local agent gateway that validates, authorises, and proxies agent tool calls.
    Run on localhost only — do NOT expose to the network without enterprise-grade
    authentication, mTLS, and policy enforcement.
servers:
  - url: http://127.0.0.1:8080
    description: Local-only default
...
paths:
...
  /tools/call:
    post:
      summary: Submit a tool call for validation and proxying
      tags: [tools]
      description: |
        Accepts a `ToolCall` payload, runs capability and schema checks, then
        returns a `ToolResult`.  When the kill-switch is active all calls are
        rejected with **503**.  If the calling user has an `allowed_tools`
        restriction, calls to tools outside that list are rejected with **403**.
      security:
        - BearerToken: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCall'
      responses:
        '200':
          description: Tool accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
        '202':
          description: Pending human approval
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: pending_approval
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  schemas:
...
    ToolCall:
      type: object
      required: [tool, args]
      additionalProperties: false
      properties:
        tool:
          type: string
        args:
          type: object
    ToolResult:
      type: object
      properties:
        tool:
          type: string
        args:
          type: object
        status:
          type: string
          example: stubbed
        message:
          type: string
        risk:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    ValidationError:
      description: Schema validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden — insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Service temporarily unavailable (kill-switch active)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Bearer token from POST /admin/login.
        Admin endpoints require the `admin` role.